% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Filter.R
\name{add_Filter}
\alias{add_Filter}
\title{Function to add a filter to a stack}
\usage{
add_Filter(object, filter)
}
\arguments{
\item{object}{\code{\linkS4class{FilterStack}} from which the filter has to be added}

\item{filter}{\code{\linkS4class{Filter}} that has to be added}
}
\description{
This function adds a filter to a stack
}
\author{
David Strebler, University of Cologne (Germany).
}

