% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Filter.R
\name{create_Filter}
\alias{create_Filter}
\title{Function to create a filter.}
\usage{
create_Filter(name, description, reference.thickness, thickness = NULL,
  reflexion, transmission)
}
\arguments{
\item{name}{\link{character}: Name of the filter.}

\item{description}{\link{character}: Description of the filter.}

\item{reference.thickness}{\link{numeric}: Reference thickness of the filter.}

\item{thickness}{\link{numeric}: Thickness of the filter (by default thickness = reference.thickness).}

\item{reflexion}{\link{numeric}: Reflexion of the filter (1-P).}

\item{transmission}{\link{numeric}: Transmission of the filter (T).}
}
\value{
This function return a new filter.
}
\description{
This function create a new filter.
}
\author{
David Strebler, University of Cologne (Germany).
}

