% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_FilterStack.R
\name{create_FilterStack}
\alias{create_FilterStack}
\title{Function to create a filter stack}
\usage{
create_FilterStack(name, description, filters)
}
\arguments{
\item{name}{\link{character}: Name of the filter stack.}

\item{description}{\link{character}: Description of the filter stack.}

\item{filters}{\link{list}: Filter included in the stack}
}
\value{
This function return a new filter stack.
}
\description{
This function create a new filter stack
}
\examples{
#Data
name <- "example"
description <- "non realistic filterStack"

filters <- default_Filters(c('example','example2'))

#Filterstack
filterstack <- create_FilterStack(name,description,filters)

plot_FilterStack(filterstack)

}
\author{
David Strebler, University of Cologne (Germany).
}

