% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Stimulation.R
\name{create_Stimulation}
\alias{create_Stimulation}
\title{Function to create a Stimulation.}
\usage{
create_Stimulation(name, description, emission)
}
\arguments{
\item{name}{\link{character}: Name of the Stimulation.}

\item{description}{\link{character}: Description of the Stimulation.}

\item{emission}{\link{numeric}: emission of the Stimulation [u.a].}
}
\value{
This function return a new Stimulation.
}
\description{
This function create a new Stimulation.
}
\author{
David Strebler, University of Cologne (Germany).
}

