% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_Material.R
\name{default_Material}
\alias{default_Material}
\title{Fonction to upload the default Material.}
\usage{
default_Material(name)
}
\arguments{
\item{name}{\link{character} (with default): name of the Material to import.}
}
\value{
This function return a list containing the Material included in the package.
}
\description{
This fonction generates a list containing the Material included in the packages
}
\examples{
example <- default_Material('example')

plot_Material(example)

}
\author{
David Strebler, University of Cologne (Germany).
}

