% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_PMT.R
\name{default_PMT}
\alias{default_PMT}
\title{Fonction to upload the default PMT.}
\usage{
default_PMT(name)
}
\arguments{
\item{name}{\link{character} (with default): name of the PMT to import.}
}
\value{
This function return a list containing the PMT included in the package.
}
\description{
This fonction generates a list containing the PMT included in the packages
}
\author{
David Strebler, University of Cologne (Germany).
}

