% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_Stimulation.R
\name{default_Stimulation}
\alias{default_Stimulation}
\title{Fonction to upload the default stimulation.}
\usage{
default_Stimulation(names)
}
\arguments{
\item{names}{\link{character} (with default): Names of the stimulation to import.}
}
\value{
This function return a list containing the stimulation included in the package.
}
\description{
This fonction generates a list containing the stimulation included in the packages
}
\author{
David Strebler, University of Cologne (Germany).
}

