% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment-class.R
\docType{class}
\name{Experiment-class}
\alias{Experiment-class}
\alias{getExperiment,Experiment-method}
\alias{setExperiment,Experiment-method}
\alias{show,Experiment-method}
\title{Class \code{Experiment}}
\description{
Object class containing the properties of aPhotomultiplier tube.
}
\author{
David Strebler
}
\keyword{classes}

