% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_PMT.R
\name{export_PMT}
\alias{export_PMT}
\title{Function to export a PMT}
\usage{
export_PMT(object, file.name)
}
\arguments{
\item{object}{\code{\linkS4class{PMT}} to export}

\item{file.name}{\link{character} name of the .FLT file that will contain the PMT properties.}
}
\value{
The function creates a new txt file containing the PMT properties.
}
\description{
This function exports a PMT as a .PMT file.
}
\examples{
example <- default_PMT('example')
file.name <- 'example'

\dontrun{
export_PMT(example, file.name)  # uncomment
}

# There is now an 'example.PMT' file in the 'working directory'.
# This file is a classical .txt file despite the extension.

file <- paste(getwd(),'/', file.name,'.PMT', sep="")
print(file)
# readLines(file)

}
\author{
David Strebler, University of Cologne (Germany).
}

