% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_Stimulation.R
\name{export_Stimulation}
\alias{export_Stimulation}
\title{Function to export a Stimulation}
\usage{
export_Stimulation(object, file.name)
}
\arguments{
\item{object}{\code{\linkS4class{Stimulation}} to export}

\item{file.name}{\link{character} name of the .FLT file that will contain the Stimulation properties.}
}
\value{
The function creates a new txt file containing the Stimulation properties.
}
\description{
This function exports a Stimulation as a .STI file.
}
\examples{
example <- default_Stimulation('example')
file.name <- 'example'

\dontrun{
export_Stimulation(example, file.name) # uncomment
}

# There is now an 'example.EXI' file in the 'working directory'.
# This file is a classical .txt file despite the extension.

file <- paste(getwd(),'/', file.name,'.EXI', sep="")
print(file)
# readLines(file)

}
\author{
David Strebler, University of Cologne (Germany).
}

