% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStack-class.R
\docType{class}
\name{FilterStack-class}
\alias{FilterStack-class}
\alias{getFilterStack}
\alias{getFilterStack,FilterStack-method}
\alias{setFilterStack}
\alias{setFilterStack,FilterStack-method}
\alias{show,FilterStack-method}
\title{Class \code{FilterStack}}
\description{
Object class containing a combination of filters.
}
\author{
David Strebler
}
\keyword{classes}

