% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Filter.R
\name{import_Filter}
\alias{import_Filter}
\title{Function to import a filter}
\usage{
import_Filter(file.name)
}
\arguments{
\item{file.name}{\link{character} name of the .FLT file containing the filter properties.}
}
\value{
The function creates a new \code{\linkS4class{Filter}} object.
}
\description{
This function import the properties of a filter previously saved in a .FLT file.
}
\author{
David Strebler, University of Cologne (Germany).
}

