% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Material.R
\name{import_Material}
\alias{import_Material}
\title{Function to import a Material}
\usage{
import_Material(file.name)
}
\arguments{
\item{file.name}{\link{character} name of the .FLT file containing the Material properties.}
}
\value{
The function creates a new \code{\linkS4class{Material}} object.
}
\description{
This function import the properties of a Material previously saved in a .Material file.
}
\author{
David Strebler, University of Cologne (Germany).
}

