% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Stimulation.R
\name{import_Stimulation}
\alias{import_Stimulation}
\title{Function to import a Stimulation}
\usage{
import_Stimulation(file.name)
}
\arguments{
\item{file.name}{\link{character} name of the .FLT file containing the Stimulation properties.}
}
\value{
The function creates a new \code{\linkS4class{Stimulation}} object.
}
\description{
This function import the properties of a Stimulation previously saved in a .EXI file.
}
\author{
David Strebler, University of Cologne (Germany).
}

