% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LumReader-package.R
\docType{package}
\name{LumReader-package}
\alias{LumReader}
\alias{LumReader-package}
\title{Tools to simulate a TL/OSL Reader}
\description{
A series of function to estimate the detection windows of a TL/OSL reader based on the filters and the PMT selected.
It also allows to check if the stimulation unit do not overlap with this detection windows.
Finally, for those only interested by filter combination, it could provide the transmission property of a filter stack.
The user can used the filters, PMT and stimulation unit already includes in the package or creates its own.
}
\details{
\tabular{ll}{  Package: \tab TLdating \cr
               Type: \tab Package\cr
               Version: \tab 0.1.0 \cr
               Date: \tab 2016-03-01 \cr
               License: \tab GPL-3
            }
}
\author{
\bold{Authors} \cr
 \tabular{ll}{  David Strebler, \tab University of Cologne, Germany
 }


 \bold{Supervisor} \cr
 \tabular{ll}{  Helmut Brückner, \tab University of Cologne, Germany \cr
                Dominik Brill, \tab University of Cologne, Germany
 }




 \bold{Project source code repository} \cr
 \url{https://github.com/dstreble/LumReader} \cr


 \bold{Package maintainer} \cr
 David Strebler, Geographisches Institut, Universitat zu Koeln, Cologne, Germany. \cr
 \email{david.strebler@uni-koeln.de}

 \bold{Acknowledgement} \cr
 This project is realized in the context of the CRC 806 “Our Way to Europe” (\url{http://www.sfb806.uni-koeln.de/}) which is funded by the German Research foundation (DFG). \cr
}
\keyword{package}

