% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Experiment.R
\name{plot_Experiment}
\alias{plot_Experiment}
\title{Function to plot a Experiment}
\usage{
plot_Experiment(object)
}
\arguments{
\item{object}{\code{\linkS4class{Experiment}} to plot}
}
\description{
This function plots the selected Experiment.
}
\examples{
# Create info
name <- 'example'
description <- 'example'

# Create reader components
filter <- default_Filters('example')
filterStack <- create_FilterStack(name, description, filter)
stimulation <- default_Stimulation('example')
PMT <- default_PMT('example')

# Create reader
reader <- create_Reader(name, description, stimulation, filterStack, PMT)

# Create material
material <- default_Material('example')

experiment <- create_Experiment(name, description,reader,material, 'OSL')

plot_Experiment(experiment)

}
\author{
David Strebler, University of Cologne (Germany).
}

