% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FilterStack.R
\name{plot_FilterStack}
\alias{plot_FilterStack}
\title{Function to plot a filter stack.}
\usage{
plot_FilterStack(object)
}
\arguments{
\item{object}{\code{\linkS4class{FilterStack}} to plot.}
}
\description{
This function plots a filter stack.
}
\examples{

#Data
name <- "example"
description <- "non realistic filterStack"

filters <- default_Filters(c('example','example2'))

#Filterstack
filterstack <- create_FilterStack(name,description,filters)

plot_FilterStack(filterstack)

}
\author{
David Strebler, University of Cologne (Germany).
}

