% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Material.R
\name{plot_Material}
\alias{plot_Material}
\title{Function to plot a Material}
\usage{
plot_Material(object)
}
\arguments{
\item{object}{\code{\linkS4class{Material}} to plot}
}
\description{
This function plots the selected Material.
}
\author{
David Strebler, University of Cologne (Germany).
}

