% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PMT.R
\name{plot_PMT}
\alias{plot_PMT}
\title{Function to plot a PMT}
\usage{
plot_PMT(object)
}
\arguments{
\item{object}{\code{\linkS4class{PMT}} to plot}
}
\description{
This function plots the selected PMT.
}
\examples{
PMT <- default_PMT('example')

plot_PMT(PMT)

}
\author{
David Strebler, University of Cologne (Germany).
}

