% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Reader.R
\name{plot_Reader}
\alias{plot_Reader}
\title{Function to plot a filter stack.}
\usage{
plot_Reader(object)
}
\arguments{
\item{object}{\code{\linkS4class{Reader}} to plot.}
}
\description{
This function plots a filter stack.
}
\author{
David Strebler, University of Cologne (Germany).
}

