% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Stimulation.R
\name{plot_Stimulation}
\alias{plot_Stimulation}
\title{Function to plot a Stimulation}
\usage{
plot_Stimulation(object)
}
\arguments{
\item{object}{\code{\linkS4class{Stimulation}} to plot}
}
\description{
This function plots the selected Stimulation.
}
\examples{
stimulation <- default_Stimulation('example')

plot_Stimulation(stimulation)


}
\author{
David Strebler, University of Cologne (Germany).
}

