% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMT-class.R
\docType{class}
\name{PMT-class}
\alias{PMT-class}
\alias{getPMT,PMT-method}
\alias{setPMT,PMT-method}
\alias{show,PMT-method}
\title{Class \code{PMT}}
\description{
Object class containing the properties of aPhotomultiplier tube.
}
\author{
David Strebler
}
\keyword{classes}

