% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reader-class.R
\docType{class}
\name{Reader-class}
\alias{Reader-class}
\alias{getReader,Reader-method}
\alias{setReader,Reader-method}
\alias{show,Reader-method}
\title{Class \code{Reader}}
\description{
Object class containing the properties of aPhotomultiplier tube.
}
\author{
David Strebler
}
\keyword{classes}

