% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stimulation-class.R
\docType{class}
\name{Stimulation-class}
\alias{Stimulation-class}
\alias{getStimulation,Stimulation-method}
\alias{setStimulation,Stimulation-method}
\alias{show,Stimulation-method}
\title{Class \code{Stimulation}}
\description{
Object class containing the properties of aPhotomultiplier tube.
}
\author{
David Strebler
}
\keyword{classes}

