% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Filter.R
\name{add_Filter}
\alias{add_Filter}
\title{Function to add a filter to a stack}
\usage{
add_Filter(object, filter)
}
\arguments{
\item{object}{\code{\linkS4class{FilterStack}} from which the filter has to be added}

\item{filter}{\code{\linkS4class{Filter}} that has to be added}
}
\description{
This function adds a filter to a stack
}
\examples{
#create filters
filter1 <- default_Filters('example')[[1]]
filter2 <- default_Filters('example2')[[1]]

#Create filterStack
name <- "example"
description <- "non realistic filterStack"

filters <- list(filter1)
filterStack <- create_FilterStack(name, description, filters)

plot_FilterStack(filterStack)

#add_Filter
filterStack <- add_Filter(filterStack, filter2)

plot_FilterStack(filterStack)

}
\author{
David Strebler, University of Cologne (Germany).
}

