% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ShinyLumReader.R
\name{call_ShinyLumReader}
\alias{call_ShinyLumReader}
\title{ShinyLumReader application}
\usage{
call_ShinyLumReader()
}
\description{
This function calls the 'shinyLumReader' application.
}
\examples{
\dontrun{
call_ShinyLumReader()
}
}
\author{
David Strebler, University of Cologne (Germany).
}

