% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_Filters.R
\name{combine_Filters}
\alias{combine_Filters}
\title{Function to combine two filters}
\usage{
combine_Filters(filter1, filter2)
}
\arguments{
\item{filter1}{\code{\linkS4class{Filter}} First of the two filters to combine.}

\item{filter2}{\code{\linkS4class{Filter}} Second of the two filters to combine.}
}
\value{
The function creates a new \code{\linkS4class{Filter}} object.
}
\description{
This function creates a new filter out of 2 filters.
}
\examples{

filter

#create filters
filter1 <- default_Filters('example')[[1]]
filter2 <- default_Filters('example2')[[1]]

new.filter <- combine_Filters (filter1, filter2)

plot_Filter(new.filter)

}
\author{
David Strebler, University of Cologne (Germany).
}

