% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Material.R
\name{create_Material}
\alias{create_Material}
\title{Function to create a Material.}
\usage{
create_Material(name, description.TL, description.OSL, TL, OSL)
}
\arguments{
\item{name}{\link{character}: Name of the material.}

\item{description.TL}{\link{character}: description of the material TL properties.}

\item{description.OSL}{\link{character}: description of the material OSL properties.}

\item{TL}{\link{numeric}: TL response of the Material [a.u.].}

\item{OSL}{\link{numeric}: TL of the Material [a.u.].}
}
\value{
This function return a new Material.
}
\description{
This function create a new Material.
}
\author{
David Strebler, University of Cologne (Germany).
}

