% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PMT.R
\name{create_PMT}
\alias{create_PMT}
\title{Function to create a PMT.}
\usage{
create_PMT(name, description, efficiency)
}
\arguments{
\item{name}{\link{character}: Name of the PMT.}

\item{description}{\link{character}: Description of the PMT.}

\item{efficiency}{\link{numeric}: efficiency of the PMT [mA/W].}
}
\value{
This function return a new PMT.
}
\description{
This function create a new PMT.
}
\author{
David Strebler, University of Cologne (Germany).
}

