% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_Filters.R
\name{default_Filters}
\alias{default_Filters}
\title{Fonction to upload the default filters.}
\usage{
default_Filters(names)
}
\arguments{
\item{names}{\link{character} (with default): Names of the filters to import.}
}
\value{
This function return a list containing the filters included in the package.
}
\description{
This fonction generates a list containing the filters included in the packages
}
\author{
David Strebler, University of Cologne (Germany).
}

