% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_Stimulation.R
\name{default_Stimulation}
\alias{default_Stimulation}
\title{Fonction to upload the default Stimulation.}
\usage{
default_Stimulation(name)
}
\arguments{
\item{name}{\link{character} (with default): name of the Stimulation to import.}
}
\value{
This function return a list containing the Stimulation included in the package.
}
\description{
This fonction generates a list containing the Stimulation included in the packages
}
\examples{
example <- default_Stimulation('example')

plot_Stimulation(example)

}
\author{
David Strebler, University of Cologne (Germany).
}

