% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_Material.R
\name{export_Material}
\alias{export_Material}
\title{Function to export a Material}
\usage{
export_Material(object, file.name)
}
\arguments{
\item{object}{\code{\linkS4class{Material}} to export}

\item{file.name}{\link{character} name of the .FLT file that will contain the Material properties.}
}
\value{
The function creates a new txt file containing the Material properties.
}
\description{
This function exports a Material as a .Material file.
}
\author{
David Strebler, University of Cologne (Germany).
}

