% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_Stimulation.R
\name{export_Stimulation}
\alias{export_Stimulation}
\title{Function to export a Stimulation}
\usage{
export_Stimulation(object, file.name)
}
\arguments{
\item{object}{\code{\linkS4class{Stimulation}} to export}

\item{file.name}{\link{character} name of the .FLT file that will contain the Stimulation properties.}
}
\value{
The function creates a new txt file containing the Stimulation properties.
}
\description{
This function exports a Stimulation as a .STI file.
}
\author{
David Strebler, University of Cologne (Germany).
}

