% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter-class.R
\docType{class}
\name{Filter-class}
\alias{Filter-class}
\alias{getFilter,Filter-method}
\alias{setFilter,Filter-method}
\alias{show,Filter-method}
\title{Class \code{Filter}}
\description{
Object class containing the properties of a optical filter.
}
\author{
David Strebler
}
\keyword{classes}

