% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Material-class.R
\docType{class}
\name{Material-class}
\alias{Material-class}
\alias{getMaterial,Material-method}
\alias{setMaterial,Material-method}
\alias{show,Material-method}
\title{Class \code{Material}}
\description{
Object class containing the properties of aPhotomultiplier tube.
}
\author{
David Strebler
}
\keyword{classes}

