% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Experiment.R
\name{plot_Experiment}
\alias{plot_Experiment}
\title{Function to plot a Experiment}
\usage{
plot_Experiment(object)
}
\arguments{
\item{object}{\code{\linkS4class{Experiment}} to plot}
}
\description{
This function plots the selected Experiment.
}
\author{
David Strebler, University of Cologne (Germany).
}

