% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Filter.R
\name{plot_Filter}
\alias{plot_Filter}
\title{Function to plot a filter}
\usage{
plot_Filter(object)
}
\arguments{
\item{object}{\code{\linkS4class{Filter}} to plot}
}
\description{
This function plots the selected filter.
}
\examples{
filters <- default_Filters(c('example', 'example2'))

plot_Filter(filters[[1]])

}
\author{
David Strebler, University of Cologne (Germany).
}

