% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_Filter.R
\name{remove_Filter}
\alias{remove_Filter}
\title{Function to remove a filter from a stack}
\usage{
remove_Filter(object, filter)
}
\arguments{
\item{object}{\code{\linkS4class{FilterStack}} from which the filter has to be removed.}

\item{filter}{\code{\linkS4class{Filter}} that has to be removed.}
}
\description{
This function removes a filter from a stack
}
\author{
David Strebler, University of Cologne (Germany).
}

